+++ actions/api-action/exhibition-actions.ts
@@ -1,42 +1,3 @@
-type IExhibitionDetailInfo = {
-  id: string;
-  ownerId: string;
-  title: string;
-  description: string;
-  location: string;
-  isParkingAvailable: boolean;
-  fee: number;
-  startDate: Date;
-  endDate: Date;
-  openTimes: string;
-  latitude: number;
-  longitude: number;
-  publicTag: string;
-};
-
-type IExhibitionFormData = {
-  title: string;
-  description: string;
-  location: string; //주소
-  isParkingAvailable: boolean;
-  fee: number;
-  startDate: Date;
-  endDate: Date;
-  openTimes: string;
-  latitude: number;
-  longitude: number;
-  publicTag: string;
-  tags: string[];
-};
-
-type IExhibitionListResponse = {
-  id: string;
-  title: string;
-  location: string;
-  startDate: Date;
-  endDate: Date;
-};
-
 const getExhibitionInfo = async (
   id: string,
 ): Promise<IExhibitionDetailInfo | Error> => {
