+++ actions/api-action/popup-actions.ts
@@ -1,42 +1,3 @@
-type IPopupDetailInfo = {
-  id: string;
-  ownerId: string;
-  title: string;
-  description: string;
-  location: string;
-  isParkingAvailable: boolean;
-  fee: number;
-  startDate: Date;
-  endDate: Date;
-  openTimes: string;
-  latitude: number;
-  longitude: number;
-  publicTag: IPublicTag;
-};
-
-type IPopupFormData = {
-  title: string;
-  description: string;
-  location: string; //주소
-  isParkingAvailable: boolean;
-  fee: number;
-  startDate: Date;
-  endDate: Date;
-  openTimes: string;
-  latitude: number;
-  longitude: number;
-  publicTag: string;
-  tags: string[];
-};
-
-type IPopupListResponse = {
-  id: string;
-  title: string;
-  location: string;
-  startDate: Date;
-  endDate: Date;
-};
-
 const getPopupInfo = async (id: string): Promise<IPopupDetailInfo | Error> => {
   try {
     const res = await fetch(`${process.env.NEXT_PUBLIC_ENDPOINT}/popup/${id}`);
