--- constants/region.ts
@@ -1,8 +1,0 @@
-export const REGIONS: ISearchRegion[] = [
-  "서울특별시",
-  "경기도",
-  "인천광역시",
-  "강원도",
-  "부산광역시",
-  "대전광역시",
-];
