--- constants/tag.ts
@@ -1,19 +1,0 @@
-export const PUBLIC_TAGS: IPublicTag[] = [
-  "브랜드",
-  "패션",
-  "뷰티",
-  "음식",
-  "홈",
-  "완구류",
-  "레저",
-  "서적",
-  "음악",
-  "펫",
-  "운동",
-  "디지털",
-  "예술",
-  "캐릭터",
-  "굿즈",
-  "전시",
-  "기타",
-];
